Small refactor for codegen
